+++ libs/utils/analytics.md
@@ -1,0 +1,25 @@
+# Adobe analytics
+
+> Add tracking attributes to the DOM
+
+## Link hierarchy attributes
+
+### `daa-im`
+
+Simple attribute flag if set to true will capture impressions for each link in that container.
+
+### `daa-lh`
+
+Hierarchy values that are used to create a complete view of where each interaction is located. 
+
+Our code will combine hierarchy values to create where (unique identifier) for every interaction.
+
+## CTA attributes
+
+(CTA English value-CTA number in current box)
+
+### `daa-ll`
+
+Interaction identifier. Such as "Learn more" for a learn more link.
+
+[Home Page Link Tracking](https://wiki.corp.adobe.com/display/~cwest/PRD%3A+Home+Page+Link+Tracking)
