+++ test/libs/utils/decorate.test.js
@@ -1,8 +1,9 @@
 /* eslint-disable no-unused-expressions */
 /* global it */
 import { expect } from '@esm-bundle/chai';
-import {isHexColorDark, getIconLibrary, initIcons, loadTokens} from '../../../libs/utils/decorate.js';
-import {readFile} from "@web/test-runner-commands";
+import { isHexColorDark, initIcons } from '../../../libs/utils/decorate.js';
+import { loadTokens } from '../../../libs/utils/utils.js';
+import { readFile } from "@web/test-runner-commands";
 
 document.head.innerHTML = await readFile({ path: './mocks/head.html' });
 document.body.innerHTML = await readFile({ path: './mocks/body.html' });
